Move snapshot mixin to actions folder
