+++ oarepo_requests/services/record/components/snapshot_component.py
@@ -1,44 +1,30 @@
 from invenio_records_resources.services.records.components import ServiceComponent
-from oarepo_requests.types.record_snapshot_mixin import create_snapshot_and_possible_event
-from oarepo_runtime.datastreams.utils import get_record_service_for_record
-from invenio_requests.records.models import RequestMetadata
-from invenio_db import db
+from opensearch_dsl.query import Bool, Term, Terms
+from oarepo_requests.actions.record_snapshot_mixin import create_snapshot_and_possible_event
+from invenio_requests.proxies import current_requests_service
+from invenio_access.permissions import system_identity
+from invenio_requests.resolvers.registry import ResolverRegistry
+from uuid import UUID
+from flask import current_app
+
 
 class RecordSnapshotComponent(ServiceComponent):
 
+    def create_snapshot(self, record):
+        topic_dict = ResolverRegistry.reference_entity(record)
+        topic_type, topic_id = next(iter(topic_dict.items()))
+
+        # find request for this record
+        requests = list(current_requests_service.search(system_identity, extra_filter=Bool(
+            must=[Term(**{f"topic.{topic_type}": topic_id}), Terms(type=current_app.config['PUBLISH_REQUEST_TYPES'])]),
+                                                        sort='newest', size=1).hits)
+
+        if requests:
+            create_snapshot_and_possible_event(record, record['metadata'], UUID(requests[0]['id']))
+
     def update(self, identity, *, record, **kwargs):
         """Update handler."""
-
-        # get all requests with record id
-        requests = [result for result in db.session.query(RequestMetadata).all() 
-            if 'topic' in result.json and any(value == record['id'] for value in result.json['topic'].values())]
-        
-        # sort to get the latest one
-        sorted_requests = sorted(requests, key=lambda x: x.updated)
-        
-        if sorted_requests:
-            service = get_record_service_for_record(record)
-            if record.is_draft:
-                record_item = service.read_draft(identity, record.pid.pid_value)
-            else:
-                record_item = service.read(identity, record.pid.pid_value)
-            
-            create_snapshot_and_possible_event(record, record_item, sorted_requests[0].id)
+        self.create_snapshot(record)
 
     def update_draft(self, identity, *, record, **kwargs):
-
-        # get all requests with record id
-        requests = [result for result in db.session.query(RequestMetadata).all() 
-            if 'topic' in result.json and any(value == record['id'] for value in result.json['topic'].values())]
-        
-        # sort to get the latest one
-        sorted_requests = sorted(requests, key=lambda x: x.updated)
-
-        if sorted_requests:
-            service = get_record_service_for_record(record)
-            if record.is_draft:
-                record_item = service.read_draft(identity, record.pid.pid_value)
-            else:
-                record_item = service.read(identity, record.pid.pid_value)
-            
-            create_snapshot_and_possible_event(record, record_item, sorted_requests[0].id)
+        self.create_snapshot(record)
